Make an array index name one element or none
Two defects met in the same arithmetic, and both ran without complaint while answering
wrongly. They are now one helper, `index_offset`, used by all four index sites.

**An out-of-range index wrapped onto a valid element.** Both parse sites narrowed the
literal with `as i32`, so on `["safe", "other"]`:

    Items[4294967295] == "other"    PASS   (wrapped to -1)
    Items[4294967296] == "safe"     PASS   (wrapped to 0)
    Items[9223372036854775807] == "other"   PASS

Reported by a reviewer with those cases. The index is now carried at the width the parser
reads it, and the bounds check does the rejecting -- retrieval already reports an
out-of-range index as unresolved, so nothing new was needed beyond not throwing the
information away. `usize::try_from` rather than `as usize` at the point of use, because on
a 32-bit target the cast would truncate and could land back inside the collection.

**A negative index was its own magnitude rather than an offset from the end.** On
`[a, b, c]`, `Items[-1]` was `b`, `Items[-2]` was `c`, and `Items[-3]` was reported out of
bounds -- inverted and off by one at once. `-1` is now the last element and `-len` the
first. Nothing asserted this in either direction and it is undocumented, which is how it
survived; `docs/CLAUSES.md` now describes both, since a reader had no way to know what
`[-1]` meant.

The `unsigned_abs` from `7cefd51` is kept and its reason restated in the helper: negating
`i64::MIN` is not representable, so the obvious spelling panics in a debug build and wraps
in release.

**`substring` narrowed its bounds to 16 bits.** `usize::from(n as u16)` meant
`substring(s, 65536, 65539)` read from 0 and returned the first three characters, and a
negative bound wrapped to a large one. Both now reject, with a message naming the argument
and the value. The function's own bounds check was already correct -- it declines and
returns nothing when the range is outside the string -- so only the conversion was wrong.

Found while checking the reviewer's index report for other instances of the same shape,
which is the argument for reading a finding as a class rather than as a case.

+++ guard/src/rules/exprs.rs
@@ -68,7 +68,13 @@ pub(crate) enum QueryPart<'loc> {
     MapKeyFilter(Option<String>, MapKeyFilterClause<'loc>),
     AllValues(Option<String>),
     AllIndices(Option<String>),
-    Index(i32),
+    /// An array index as written in the rule, kept at the width the parser reads.
+    ///
+    /// `i32` before, narrowed with `as i32` at both parse sites, which wrapped instead of rejecting:
+    /// `Items[4294967296]` became `Items[0]` and the clause then compared the wrong element and
+    /// passed. Retrieval already reports an out-of-range index as unresolved, so widening is all that
+    /// is needed -- the bounds check does the rejecting.
+    Index(i64),
     Filter(Option<String>, Conjunctions<GuardClause<'loc>>),
 }
